refactor: Change the path of 'SemiLagrangian' related files to 'Solver'
