+++ Sources/Core/Solver/Grid/GridFluidSolver2.cpp
@@ -9,7 +9,7 @@
 // property of any third parties.
 
 #include <Core/Array/ArrayUtils.hpp>
-#include <Core/SemiLagrangian/CubicSemiLagrangian2.hpp>
+#include <Core/Solver/Advection/CubicSemiLagrangian2.hpp>
 #include <Core/Solver/Grid/GridBackwardEulerDiffusionSolver2.hpp>
 #include <Core/Solver/Grid/GridFluidSolver2.hpp>
 #include <Core/Solver/Grid/GridFractionalSinglePhasePressureSolver2.hpp>
