+++ Tests/ManualTests/GridSmokeSolverTests.cpp
@@ -8,7 +8,7 @@
 #include <Core/Geometry/RigidBodyCollider3.hpp>
 #include <Core/Geometry/Sphere2.hpp>
 #include <Core/Geometry/Sphere3.hpp>
-#include <Core/SemiLagrangian/CubicSemiLagrangian3.hpp>
+#include <Core/Solver/Advection/CubicSemiLagrangian3.hpp>
 #include <Core/Solver/Grid/GridSinglePhasePressureSolver2.hpp>
 #include <Core/Solver/Grid/GridSmokeSolver2.hpp>
 #include <Core/Solver/Grid/GridSmokeSolver3.hpp>
