+++ Tests/ManualTests/IterativeLevelSetSolversTests.cpp
@@ -3,7 +3,7 @@
 #include <Core/Field/CustomVectorField2.hpp>
 #include <Core/Grid/CellCenteredScalarGrid2.hpp>
 #include <Core/Grid/CellCenteredScalarGrid3.hpp>
-#include <Core/SemiLagrangian/CubicSemiLagrangian2.hpp>
+#include <Core/Solver/Advection/CubicSemiLagrangian2.hpp>
 #include <Core/Solver/LevelSet/ENOLevelSetSolver2.hpp>
 #include <Core/Solver/LevelSet/ENOLevelSetSolver3.hpp>
 #include <Core/Solver/LevelSet/UpwindLevelSetSolver2.hpp>
